+++ packages/feedback/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./diagnostics.js";
 export * from "./draft.js";
 export * from "./privacy.js";
+export * from "./submission.js";